:recycle: Refactor : Image 삭제시 Validation 추가 <body> - removeImage()에서 해당 스케쥴에 대한 이미지가 맞는지 확인 로직 추가
<footer>
- 관련 : #174

+++ application/external-api/src/main/java/com/namo/spring/application/external/api/individual/service/ImageService.java
@@ -48,7 +48,10 @@ public class ImageService {
 		);
 	}
 
-	public void removeImage(Image img) {
+	public void removeImage(Long scheduleId, Image img) {
+		if (!img.getSchedule().getId().equals(scheduleId)) {
+			throw new IndividualException(ErrorStatus.NOT_IMAGE_IN_DIARY);
+		}
 		imageRepository.delete(img);
 	}
 }
